Add new project and new ontology wizards
The new project wizard configures the project to have OML and Xtext natures and creates a src directory.

The new ontology wizard creates a vocabulary, bundle, or description ontology, and allows the user to specify a namespace IRI, separator, and prefix.

By default, description ontologies are saved as OML XMI and vocabulary and bundle ontologies are saved as OML text.